+++ R/00-Namespace.R
@@ -15,6 +15,7 @@
 #' @rawNamespace import(shiny)
 #' @importFrom rjson fromJSON
 #' @importFrom DataTools importDataUI importDataServer importOptions
+#' @importFrom shinyTools dataExportButton dataExportServer headerButtonsUI includeShinyToolsCSS
 #' @importFrom yaml read_yaml
 
 NULL