+++ inst/app/ui.R
@@ -2,6 +2,7 @@ library(shiny)
 
 tagList(
   navbarPage(
+    header = shinyTools::includeShinyToolsCSS(),
     title = paste("MapR", packageVersion("MapR")),
     theme = shinythemes::shinytheme("flatly"),
     position = "fixed-top",
@@ -12,6 +13,7 @@ tagList(
       mapPanelUI(id = "map_panel")
     )
   ),
+  shinyTools::headerButtonsUI(id = "header", help_link = "https://pandora-isomemo.github.io/MapR/articles/how-to-use-MapR.html"),
   tags$head(
     tags$link(rel = "stylesheet", type = "text/css", href = "custom.css")
 ),