+++ NEWS.md
@@ -1,3 +1,18 @@
+# MapR 24.3.0
+
+## Updates
+- applying exportData module from shinyTools package, removing the code from this package
+
+# MapR 24.2.1
+
+## New Features
+- display and download of time plot data
+
+# MapR 24.2.0
+
+## New Features
+- adds header with Pandora, IsoMemo and help links
+
 # MapR 24.1.2
 
 ## Bug Fixes